Refactor/review preference refactoring/113

+++ Readio/src/main/java/ninegle/Readio/book/service/PreferenceService.java
@@ -8,8 +8,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
+import lombok.extern.slf4j.Slf4j;
 import ninegle.Readio.book.dto.preferencedto.PreferenceResponseDto;
 import ninegle.Readio.book.dto.preferencedto.PreferenceListResponseDto;
 import ninegle.Readio.book.dto.PaginationDto;
@@ -18,6 +20,8 @@ import ninegle.Readio.book.domain.Book;
 import ninegle.Readio.book.domain.Preference;
 import ninegle.Readio.book.dto.BookIdRequestDto;
 import ninegle.Readio.book.repository.PreferencesRepository;
+import ninegle.Readio.global.exception.BusinessException;
+import ninegle.Readio.global.exception.domain.ErrorCode;
 import ninegle.Readio.global.unit.BaseResponse;
 import ninegle.Readio.user.domain.User;
 import ninegle.Readio.user.service.UserContextService;
@@ -30,6 +34,7 @@ import ninegle.Readio.user.service.UserService;
  * author:  gigol
  * purpose: 
  */
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class PreferenceService {
@@ -41,30 +46,42 @@ public class PreferenceService {
 	private final UserContextService userContextService;
 
 	public Preference getPreferenceByBookAndUser(Book book,User user){
-		return preferencesRepository.findByBookAndUser(book,user).getFirst();
+		return preferencesRepository.findPreferenceByBookAndUser(book,user)
+			.orElseThrow(() -> new BusinessException(ErrorCode.BOOK_ALREADY_IN_PREFERENCE));
 	}
 
 
-	public ResponseEntity<BaseResponse<Void>> save(BookIdRequestDto dto) {
+	@Transactional
+	public PreferenceResponseDto save(Long userId,BookIdRequestDto dto) {
 
-		Book book =bookService.getBookById(dto.getId());
-		User user = userService.getById(userContextService.getCurrentUserId());
+		Book book = bookService.getBookById(dto.getId());
+		User user = userService.getById(userId);
 
-		preferencesRepository.save(preferenceMapper.toEntity(user,book));
-		return BaseResponse.ok("관심도서 추가가 정상적으로 수행되었습니다.", null,HttpStatus.CREATED);
+
+
+
+		Preference preference = preferenceMapper.toEntity(user, book);
+		preferencesRepository.save(preference);
+		return preferenceMapper.toPreferenceDto(preference);
 	}
 
-	public ResponseEntity<BaseResponse<Void>> delete(Long bookId) {
+	@Transactional
+	public ResponseEntity<BaseResponse<Void>> delete(Long userId,Long bookId) {
 		Book book =bookService.getBookById(bookId);
-		User user = userService.getById(userContextService.getCurrentUserId());
+		User user = userService.getById(userId);
 		Preference preference = getPreferenceByBookAndUser(book, user);
+
+		//404 Not Found
+		if(preference==null) {
+			throw new BusinessException(ErrorCode.PREFERENCE_NOT_FOUND);
+		}
 
 		preferencesRepository.delete(preference);
 		return BaseResponse.ok("삭제가 성공적으로 수행되었습니다.",null,HttpStatus.OK);
 	}
 
-	public ResponseEntity<BaseResponse<PreferenceListResponseDto>> getPreferenceList(int page, int size) {
-		User user = userService.getById(userContextService.getCurrentUserId());
+	public PreferenceListResponseDto getPreferenceList(Long userId,int page, int size) {
+		User user = userService.getById(userId);
 		Pageable pageable = PageRequest.of(page-1,size);
 		long count = preferencesRepository.countByUser(user);
 
@@ -72,8 +89,7 @@ public class PreferenceService {
 		List<PreferenceResponseDto> preferenceList = preferenceMapper.toPreferenceDto(preferences);
 
 		PaginationDto paginationDto = preferenceMapper.toPaginationDto(count, page, size);
-		PreferenceListResponseDto preferenceListDto = preferenceMapper.toPreferenceListDto(preferenceList, paginationDto);
 
-		return BaseResponse.ok("관심도서 조회가 정상적으로 수행되었습니다.",preferenceListDto,HttpStatus.OK);
+		return preferenceMapper.toPreferenceListDto(preferenceList, paginationDto);
 	}
 }
